mdbook: Update background examples for `mdbook test`
The examples are updated to pass:
`rustup run nightly-2025-09-19 mdbook test docs -L target\debug\deps`

Most code snippets are marked `no_run` as exercising the full Patina
code does not make sense in the context of the example. A few are
marked `ignore` if compilation would have a lot of overhead with
little return.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>