Add a HOB list pointer to Core storage
The HOB list is relocated to DXE allocated memory shortly into
DXE core entry. The Rust allocated HOB list is not binary
compatible with the C HOB list. The Rust HOB list is a vector of
references to HOB structures while the C hob list is a compact
series of binary HOB structures.

While the Rust HOB list is sufficient for most use cases that simply
need information from the HOBs in a more type safe and ergonomic
interface, there are still use cases that require binary
compatibility of the HOB list such as placement of the HOB into the
system table where it is discovered by other modules, including
C modules.

Today, this works because the HOB list pointer is direclty passed
to the system table initialization function where it is placed into
the HOB list system table. This change updates the stored HOB list
context to maintain a C HOB list pointer in addition to the Rust
managed list so all HOB list context is available in the structure.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>