patina_dxe_core/fv: Fix `test_fv_functionality()` panic
Fixes #1127

Code is currently commented out in `test_fv_functionality()` that
calls `fv_read_file()` with a non-null buffer and buffer size of
zero.

This change updates `fv_read_file()` to follow the same behavior
as the C DXE Core `FvReadFile()` function in FwVolRead.c that will
return `BUFFER_TOO_SMALL` in this scenario.

Another difference in behavior is that if the caller provides a
non-null buffer but a buffer size less than the file content size,
the C implementation will truncate the copy size to the size provided
and perform the copy with that size. That change is made in a separate
commit.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>